prevent infinite recursion when user coded anything wrong in a custom assert.pushResult() override, as shown, f.e., in the last test in unhandled-rejection.js test files.
Note that this is a bit tricky to accomplish safely as, for instance, simply throwing an exception once we've detected such a callstack run-away will very probably re-start the same problem from scratch as the onError handler will want to report the failure it 'caught' just now. Hence we make sure the handling code has a *very small* risk of re-triggering/looping indefinitely.

We count the nesting levels and when it gets obvious, we cop out by invoking the hardcoded base method.